refactor: moving the container group ports schema into containers
